server/requirements.txt
flask==3.0.2
pillow==10.4.0
numpy==1.26.4
tensorflow-cpu==2.15.0